% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites-package.R
\docType{package}
\name{cites-package}
\alias{cites}
\alias{cites-package}
\title{cites: Fast access to the  CITES Wildlife Trade Database}
\description{
This package provides the full extracted data and metadata from the CITES
wildlife trade database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliance/cites#readme}
  \item \url{https://trade.cites.org/}
  \item Report bugs at \url{https://github.com/ecohealthalliance/cites/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{ross@ecohealthalliance.org}

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder]
}

}
\keyword{internal}
