% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{cites_metadata}
\alias{cites_metadata}
\alias{metadata}
\title{CITES Field Descriptions}
\usage{
cites_metadata()
}
\value{
A tibble with field, code, and code description
}
\description{
This function returns a data frame field descriptions for \code{\link[=cites_data]{cites_data()}}.
}
\details{
This information is taken from
\href{https://github.com/ecohealthalliance/cites/tree/master/inst/extdata}{"A guide to using the CITES Trade Database"},
from the CITES websites.

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- cites:::tabular(cites::cites_metadata())
    } else {
    tmp <- tempfile(fileext=".html")
      htmlwidgets::saveWidget(DT::datatable(cites::cites_metadata(), rownames = FALSE, width=700), tmp)
      mytext <- paste('Below is a searchable version of the CITES field descriptions',
      '\\\out{<div style="width:100\%">',
         paste(stringi::stri_subset_regex(readLines(tmp), "^</?(!DOCTYPE|meta|body|html)",negate=TRUE), collapse="\n"),
      '</div>}',
      sep="\n")
    }
    mytext
  }
}

\if{text,latex}{ \Sexpr[echo=FALSE, results=rd, stage=build]{cites:::tabular(cites::cites_metadata())}}
}
\seealso{
\code{\link[=cites_codes]{cites_codes()}} \code{\link[=cites_data]{cites_data()}} \code{\link[=cites_parties]{cites_parties()}}
}
