% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{cites_parties}
\alias{cites_parties}
\alias{parties}
\title{Parties to the CITES treaty.}
\usage{
cites_parties()
}
\value{
A tibble
}
\description{
This function returns a data frame witha list of countries party to CITES
and their date of joining the treaty.
}
\details{
\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- ""
    } else {
    tmp <- tempfile(fileext=".html")
      htmlwidgets::saveWidget(DT::datatable(cites::cites_parties(), rownames = FALSE, width=700), tmp)
      mytext <- paste('Below is a searchable table of CITES parties.',
      '\\\out{<div style="width:100\%">',
         paste(stringi::stri_subset_regex(readLines(tmp), "^</?(!DOCTYPE|meta|body|html)",negate=TRUE), collapse="\n"),
      '</div>}',
      sep="\n")
    }
    mytext
  }
}

\if{text,latex}{ The HTML version of this help file includes a searchable table of the CITES parties }
}
\seealso{
\code{\link[=cites_codes]{cites_codes()}} \code{\link[=cites_metadata]{cites_metadata()}}
}
