% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{cites_codes}
\alias{cites_codes}
\alias{codes}
\title{CITES Code Values}
\usage{
cites_codes()
}
\value{
A tibble with fields and descriptions
}
\description{
This function returns a data frame with descriptions of all the code values
used in \code{\link[=cites_data]{cites_data()}}.  This is useful for lookup
as well as merging with the data for more descriptive summaries.
}
\details{
These values are drawn from
\href{https://github.com/ecohealthalliance/cites/tree/master/inst/extdata}{"A guide to using the CITES Trade Database"},
from the CITES websites.

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- c('In RStudio, this help file includes a searchable table of values.')
    } else {
    tmp <- tempfile(fileext=".html")
      htmlwidgets::saveWidget(DT::datatable(cites::cites_codes(), rownames = FALSE, width=700), tmp)
      mytext <- paste('Below is a searchable table of the CITES codes.',
      '\\\out{<div style="width:100\%">',
         paste(stringi::stri_subset_regex(readLines(tmp), "^</?(!DOCTYPE|meta|body|html)",negate=TRUE), collapse="\n"),
      '</div>}\n',
      sep="\n")
    }
    mytext
  }
}

\if{text,latex}{The HTML version of this help file includes a searchable table of the CITES codes.}
}
\seealso{
\code{\link[=cites_metadata]{cites_metadata()}} \code{\link[=cites_data]{cites_data()}}
}
